FIX : 주문서 작성 기능 오류 해결, dto 컬럼 데이터형 수정

+++ src/main/java/com/capstone/dto/UserDto.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 
 @Getter
 public class UserDto {
-    private Long uId;
+    private int uId;
     private String pw;
     private String name;
     private String bank;
